perf: optimize images, defer network map, fix iOS hero viewport
- Resize + convert oversized JPEGs to WebP (~12MB -> ~1.5MB total),
  rename to about.webp / build-1..4.webp, drop unused originals
- Load the dotted network map via dynamic import, desktop-only, so
  mobile no longer builds the ~6k-node SVG or ships the dotted-map lib
- Hero minHeight 100dvh -> 100svh to stop the iOS URL-bar reflow jump
