+++ rc-website/src/components/VisionSection.tsx
@@ -1,12 +1,25 @@
 "use client";
 
-import { useId } from "react";
+import { useEffect, useState } from "react";
 import { motion } from "motion/react";
-import { DottedMap, type Marker } from "@/components/ui/dotted-map";
+import dynamic from "next/dynamic";
+import type { OrgMarker } from "./NetworkMap";
 
-type FlagOverlay = { type: "flag"; countryCode: string; label: string };
-type DotOverlay = { type: "dot"; label: string };
-type OrgMarker = Marker & { overlay: FlagOverlay | DotOverlay };
+// Loaded only on desktop, after mount — keeps the ~6k-node dotted map (and the
+// `dotted-map` lib) out of the initial bundle and off the mobile main thread.
+const NetworkMap = dynamic(() => import("./NetworkMap"), { ssr: false });
+
+function useIsDesktop() {
+  const [isDesktop, setIsDesktop] = useState(false);
+  useEffect(() => {
+    const mq = window.matchMedia("(min-width: 768px)");
+    const update = () => setIsDesktop(mq.matches);
+    update();
+    mq.addEventListener("change", update);
+    return () => mq.removeEventListener("change", update);
+  }, []);
+  return isDesktop;
+}
 
 type Org = {
   name: string;
@@ -54,7 +67,7 @@ const markers: OrgMarker[] = orgs
   });
 
 export const VisionSection = () => {
-  const id = useId();
+  const isDesktop = useIsDesktop();
 
   return (
     <section
@@ -118,85 +131,7 @@ export const VisionSection = () => {
         </p>
 
         <div className="grid grid-cols-1 lg:grid-cols-12 gap-[5vh] items-center">
-          <div className="hidden md:block lg:col-span-9 relative overflow-hidden rounded-vh-md">
-            <div
-              aria-hidden="true"
-              className="pointer-events-none absolute inset-0 z-10"
-              style={{
-                background:
-                  "radial-gradient(ellipse 75% 70% at center, transparent 65%, #0d0d0d 100%)",
-              }}
-            />
-            <DottedMap<OrgMarker>
-              markers={markers}
-              className="aspect-[2/1] lg:aspect-[20/10]"
-              renderMarkerOverlay={({ marker, x, y, r, index }) => {
-                if (marker.overlay.type === "dot") {
-                  return (
-                    <g style={{ pointerEvents: "none" }}>
-                      <circle
-                        cx={x}
-                        cy={y}
-                        r={r * 1.6}
-                        fill="rgba(255,255,255,0.18)"
-                      />
-                      <circle cx={x} cy={y} r={r} fill="#ffffff" />
-                    </g>
-                  );
-                }
-
-                const { countryCode, label } = marker.overlay;
-                const href = `https://flagcdn.com/w80/${countryCode}.webp`;
-                const clipId = `${id}-flag-${index}`.replace(/:/g, "-");
-                const imgR = r * 0.75;
-                const fontSize = r * 0.9;
-                const pillH = r * 1.5;
-                const pillW = label.length * (fontSize * 0.62) + r * 1.4;
-                const pillX = x + r + r * 0.6;
-                const pillY = y - pillH / 2;
-
-                return (
-                  <g style={{ pointerEvents: "none" }}>
-                    <clipPath id={clipId}>
-                      <circle cx={x} cy={y} r={imgR} />
-                    </clipPath>
-                    <circle
-                      cx={x}
-                      cy={y}
-                      r={imgR + 0.15}
-                      fill="rgba(255,255,255,0.9)"
-                    />
-                    <image
-                      href={href}
-                      x={x - imgR}
-                      y={y - imgR}
-                      width={imgR * 2}
-                      height={imgR * 2}
-                      preserveAspectRatio="xMidYMid slice"
-                      clipPath={`url(#${clipId})`}
-                    />
-                    <rect
-                      x={pillX}
-                      y={pillY}
-                      width={pillW}
-                      height={pillH}
-                      rx={pillH / 2}
-                      fill="rgba(0,0,0,0.65)"
-                    />
-                    <text
-                      x={pillX + r * 0.7}
-                      y={y + fontSize * 0.35}
-                      fontSize={fontSize}
-                      fill="white"
-                      fontWeight={600}
-                    >
-                      {label}
-                    </text>
-                  </g>
-                );
-              }}
-            />
-          </div>
+          {isDesktop && <NetworkMap markers={markers} />}
 
           <div className="lg:col-span-3 flex flex-col gap-[2.5vh]">
             <h3
